fix settings and creat new app users
