+++ learnsite/core/admin.py
@@ -1,5 +1,3 @@
 from django.contrib import admin
-from .models import UserLearn
 
-
-admin.site.register(UserLearn)
+# Register your models here.
